# Warm up worker processes before switching to them