While the workers are starting up, we already process assets on the main thread. But it turns out that even after the workers have started, it takes a bit of time for them to warm up. The JS engine needs to do JIT things, optimize the code of some giant libraries like babel, etc.

We now send assets to both the worker and main process simultaneously in order to warm up the workers, and then transition to the workers only once they are warmed up. This seems to cause a pretty big improvement in perf for small apps, while large apps remain pretty much the same.

**Small app (`import 'react'`):**

Without patch (with workers): 752ms
Without workers: 357ms
With patch: 487ms

**Large app:**

Without patch (with workers): 6.76s
Without workers: 15.53s
With patch: 6.77s

cc. @DeMoorJasper @brandon93s Want to do some testing?