@brandon93s 2018-01-16
👍 Seeing a consistent ~15% speed improvement [examples/Vue](https://github.com/parcel-bundler/examples/tree/master/Vue) and ~45% improvement on [examples/hyperapp](https://github.com/parcel-bundler/examples/tree/master/hyperapp) on my PC. No improvement for [examples/react](https://github.com/parcel-bundler/examples/tree/master/react) - maybe a slight slowdown if anything.