@DeMoorJasper 2018-01-16
I've tested in on react-news (which is my go to repo for the worker tests) there i noticed a difference of -0.2secs, so it appears this also speeds up medium sized application building.
On a supertiny application it took half as long as on the master branch, so about as fast as synchronous.
Seems to be working extremely well 💯